+++ public_html/wp-content/themes/pattern-directory/css/components/_components.scss
@@ -21,11 +21,10 @@
 @import "../../../wporg/css/components/wporg-footer";
 @import "../../../wporg/css/components/wporg-header";
 @import "category-context-bar";
-@import "category-menu";
-@import "category-search";
 @import "copy-button";
 @import "favorite-button";
 @import "main-navigation";
+@import "menu";
 @import "page";
 @import "pattern-grid-menu";
 @import "pattern-grid";
@@ -35,6 +34,7 @@
 @import "pattern-report-modal";
 @import "pattern";
 @import "search-form";
+@import "search";
 @import "section-heading";
 @import "site-content";
 @import "site-header";
