+++ public_html/wp-content/themes/pattern-directory/css/components/_search.scss
