+++ public_html/wp-content/themes/pattern-directory/src/components/search/index.js
